chore(test): split tests
